Fixed the copy paste system and added a test section system - read desc for how to use that.
To use the test section system, it's pretty simple. Just make a folder called Test in the root of the project, and a folder inside it with whatever name you want. Inside that, put the map files, and make sure to replace the section name in their filenames with the previously chosen name. If you want to load that, on the section selection screen type in the name and hit "Load Test Section"